Fix Casting issue

+++ WowsKarma.Common/Models/PostFlairs.cs
@@ -49,7 +49,7 @@ namespace WowsKarma.Common.Models
 
 		public static PostFlairs ToEnum(this PostFlairsParsed flairsParsed)
 		{
-			ushort flairCount = 0x00;
+			int flairCount = 0x00;
 
 			flairCount += flairsParsed.Performance is null ? 0x00 : flairsParsed.Performance.Value ? 0x01 : 0x02;
 			flairCount += flairsParsed.Teamplay is null ? 0x00 : flairsParsed.Teamplay.Value ? 0x04 : 0x08;
